Upstream fixed this test's timing while this branch had it deleted, so the merge
conflicted on a file it had removed. Restoring it was the right resolution --
what it asserts is that concurrent token-audit writes all land, which is as true
of a store module as it was of an in-process database -- and it immediately
found two defects that no other test reaches.

Its only tie to SQLite was counting rows by preparing SQL on db1_conn().
db1_token_audit_by_source() answers the same question through the client, so the
count comes back over the bus. Upstream's timing fix is untouched.

THE SQL STAGE COULD NOT CARRY A NUMBER. token_audit.estimated_cost_usd is
NUMERIC, so pgx hands back a pgtype.Numeric, and the cell encoder had no case
for it: token_audit_by_source, _by_role, _by_tool, _by_model and the spend
breakdown all failed outright with "the store cannot send a pgtype.Numeric over
the wire". Fixed at the encoder, not by casting each query, because that is
where the wire's vocabulary belongs and a cast would have to be remembered for
every NUMERIC column added later.

AND THEN THE INTEGERS WOULD HAVE GONE QUIET. SUM() over bigint also returns
NUMERIC -- PostgreSQL widens it so the total cannot overflow -- so teaching the
stage that type would have turned a loud failure into a silent one: 31
aggregates across two families would have arrived as floats where the family
declares an integer, and a float64 cannot represent every int64 exactly. Every
one of them now says ::bigint. Only columns the schema declares as integers were
touched; SUM over a money column stays NUMERIC and rides the new case.

Verified against PostgreSQL 17.11: 2000 concurrent inserts from eight threads,
zero drops, plus the async and ingress paths, all through the real store module.
The test skips honestly where there is no database.

All 63 suites that link the store fixture were run twice. Four fail, all of them
before this merge and none touching what it changed: unit-test-agent,
unit-test-curiosity, unit-test-kb-doc-pdf and unit-test-kb-mgmt-live.
unit-test-curiosity-resolve, which was one of them, builds and passes now.

+++ server-go/modules/aimee/families/agent_work_log.go
@@ -247,10 +247,10 @@ const (
 	agentLogHUDSummarySQL = `SELECT COUNT(*),
 	                                COUNT(*) FILTER (WHERE success),
 	                                COUNT(*) FILTER (WHERE NOT success),
-	                                COALESCE(SUM(prompt_tokens), 0),
-	                                COALESCE(SUM(completion_tokens), 0),
-	                                COALESCE(SUM(turns), 0),
-	                                COALESCE(SUM(tool_calls), 0),
+	                                COALESCE(SUM(prompt_tokens)::bigint, 0),
+	                                COALESCE(SUM(completion_tokens)::bigint, 0),
+	                                COALESCE(SUM(turns)::bigint, 0),
+	                                COALESCE(SUM(tool_calls)::bigint, 0),
 	                                COALESCE(AVG(latency_ms), 0),
 	                                COUNT(*) FILTER (
 	                                    WHERE created_at > now()
@@ -265,20 +265,20 @@ const (
 
 	agentLogPrometheusSQL = `SELECT agent_name, role, COUNT(*),
 	                                COUNT(*) FILTER (WHERE success),
-	                                COALESCE(SUM(prompt_tokens), 0),
-	                                COALESCE(SUM(completion_tokens), 0),
+	                                COALESCE(SUM(prompt_tokens)::bigint, 0),
+	                                COALESCE(SUM(completion_tokens)::bigint, 0),
 	                                COALESCE(AVG(latency_ms), 0),
-	                                COALESCE(SUM(tool_calls), 0)
+	                                COALESCE(SUM(tool_calls)::bigint, 0)
 	                           FROM agent_log
 	                          GROUP BY agent_name, role
 	                          ORDER BY COUNT(*) DESC, agent_name, role
 	                          LIMIT $1`
 
 	agentLogStatsSQL = `SELECT COUNT(*),
-	                           COALESCE(SUM(turns), 0),
-	                           COALESCE(SUM(tool_calls), 0),
-	                           COALESCE(SUM(prompt_tokens), 0),
-	                           COALESCE(SUM(completion_tokens), 0),
+	                           COALESCE(SUM(turns)::bigint, 0),
+	                           COALESCE(SUM(tool_calls)::bigint, 0),
+	                           COALESCE(SUM(prompt_tokens)::bigint, 0),
+	                           COALESCE(SUM(completion_tokens)::bigint, 0),
 	                           COUNT(*) FILTER (WHERE success)
 	                      FROM agent_log WHERE ` + sinceSecsFilter
 )
